Also use TransformableOps for unmerging Sierra records
