--- sierra_adapter/sierra_item_merger/src/main/scala/uk/ac/wellcome/platform/sierra_item_merger/links/ItemUnmerger.scala
@@ -1,37 +1,0 @@
-package uk.ac.wellcome.platform.sierra_item_merger.links
-
-import uk.ac.wellcome.sierra_adapter.model.{
-  SierraItemRecord,
-  SierraTransformable
-}
-
-object ItemUnmerger {
-
-  def unmergeItemRecord(
-    sierraTransformable: SierraTransformable,
-    itemRecord: SierraItemRecord): Option[SierraTransformable] = {
-    if (!itemRecord.unlinkedBibIds.contains(sierraTransformable.sierraId)) {
-      throw new RuntimeException(
-        s"Non-matching bib id ${sierraTransformable.sierraId} in item unlink bibs ${itemRecord.unlinkedBibIds}")
-    }
-
-    val itemRecords =
-      sierraTransformable.itemRecords
-        .filterNot {
-          case (id, currentItemRecord) =>
-            val matchesCurrentItemRecord = id == itemRecord.id
-
-            val modifiedAfter = itemRecord.modifiedDate.isAfter(
-              currentItemRecord.modifiedDate
-            )
-
-            matchesCurrentItemRecord && modifiedAfter
-        }
-
-    if (sierraTransformable.itemRecords != itemRecords) {
-      Some(sierraTransformable.copy(itemRecords = itemRecords))
-    } else {
-      None
-    }
-  }
-}
